% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mustache_context.R
\name{build_mustache_context}
\alias{build_mustache_context}
\title{Convert a codebook_df into a Mustache context}
\usage{
build_mustache_context(cb)
}
\arguments{
\item{cb}{A codebook_df.}
}
\value{
A list suitable for whisker::whisker.render().
}
\description{
Convert a codebook_df into a Mustache context
}
\keyword{internal}
