% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_questionnaire.R
\name{link_questionnaire}
\alias{link_questionnaire}
\title{Cross-check questionnaire and responses}
\usage{
link_questionnaire(responses, questionnaire)
}
\arguments{
\item{responses}{A dataset_df with responses.}

\item{questionnaire}{A dataset_df with questionnaire definitions.}
}
\value{
TRUE if valid, otherwise a list of mismatches.
}
\description{
Ensures variables in responses match those defined in the questionnaire.
}
