% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_variable.R
\name{link_variable}
\alias{link_variable}
\title{Attach metadata to a variable in a dataset}
\usage{
link_variable(responses, variable, variable_uri, code_list_uri = NULL)
}
\arguments{
\item{responses}{A dataset_df with responses.}

\item{variable}{The name of the variable to annotate.}

\item{variable_uri}{A URI for the variable.}

\item{code_list_uri}{Optional URI of the associated code list.}
}
\value{
The modified dataset_df.
}
\description{
Attach metadata to a variable in a dataset
}
