% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_codebook_df.R
\name{new_codebook_df}
\alias{new_codebook_df}
\title{Create a new codebook dataset}
\usage{
new_codebook_df(concepts, scheme_meta)
}
\arguments{
\item{concepts}{A data.frame with columns: concept_id, pref_label, notation,
language. Optional: position, is_missing, exact_match, definition.}

\item{scheme_meta}{A named list with metadata: doc_uri, scheme_id,
scheme_fragment (optional), version, label.}
}
\value{
An object of class \code{codebook_df}, extending \code{dataset_df}.
}
\description{
Define a codebook as a dataset_df object with metadata and concepts.
}
