% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_codebook_hash.R
\name{publish_codebook_hash}
\alias{publish_codebook_hash}
\title{Publish a codebook as HTML and Turtle}
\usage{
publish_codebook_hash(cb, out_dir, templates_dir, html = TRUE, ttl = TRUE)
}
\arguments{
\item{cb}{A codebook_df.}

\item{out_dir}{Output directory (created if missing).}

\item{templates_dir}{Path to Mustache templates.}

\item{html}{Logical, render HTML index page.}

\item{ttl}{Logical, render Turtle file.}
}
\value{
Path to output directory.
}
\description{
Uses Mustache templates to render human and machine-readable codebooks.
}
