% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_codebook_csv.R
\name{read_codebook_csv}
\alias{read_codebook_csv}
\title{Read a codebook from CSV}
\usage{
read_codebook_csv(path)
}
\arguments{
\item{path}{Path to a CSV file with codebook fields.}
}
\value{
A \code{codebook_df}.
}
\description{
Read a codebook from CSV
}
