% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slugify_id.R
\name{slugify_id}
\alias{slugify_id}
\title{Make a safe slug for URIs}
\usage{
slugify_id(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector with safe identifiers.
}
\description{
Make a safe slug for URIs
}
\keyword{internal}
