% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_categoricals.R
\name{validate_categoricals}
\alias{validate_categoricals}
\title{Validate categorical values in a response dataset}
\usage{
validate_categoricals(responses, codebooks)
}
\arguments{
\item{responses}{A dataset_df with responses.}

\item{codebooks}{A named list of codebook_df objects, keyed by scheme_uri.}
}
\value{
TRUE if valid, otherwise a list of problems.
}
\description{
Validate categorical values in a response dataset
}
