% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_codebook_df.R
\name{validate_codebook_df}
\alias{validate_codebook_df}
\title{Validate a codebook dataset}
\usage{
validate_codebook_df(cb)
}
\arguments{
\item{cb}{A \code{codebook_df}.}
}
\value{
TRUE if valid, otherwise throws an error.
}
\description{
Checks uniqueness of concept IDs and safety of URI fragments.
}
